requirements.txt
Django==2.0.7
django-appconf ==1.0.2
django-quill ==0.1.0
django-scheduler==0.8.8
django-tinymce4-lite ==1.7.3
django-filebrowser-no-grappelli==3.7.5
icalendar==4.0.2
jsmin ==2.2.2
pip ==18.1
python-dateutil==2.7.3
pytz==2018.5
setuptools==39.1.0
simplejson==3.16.0
six==1.11.0
django-bower==5.2.0
